# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Ids of the chosen superhero avatar parts.
 */
export interface AvatarConfig {
  mask: string;
  cape: string;
  body: string;
  accessory: string;
}

export interface LeaderboardEntry {
  id: number;
  nickname: string;
  coins: number;
  avatar: AvatarConfig;
  rank: number;
  updatedAt: string;
}

export interface LeaderboardEntryInput {
  /** @minLength 1 */
  deviceId: string;
  /**
     * @minLength 1
     * @maxLength 20
     */
  nickname: string;
  /** @minimum 0 */
  coins: number;
  avatar: AvatarConfig;
}

export type ListLeaderboardEntriesParams = {
/**
 * @minimum 1
 * @maximum 200
 */
limit?: number;
};