# lib/api-zod/src/generated/types/avatarConfig.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Ids of the chosen superhero avatar parts.
 */
export interface AvatarConfig {
  mask: string;
  cape: string;
  body: string;
  accessory: string;
}