# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AvatarConfig } from './avatarConfig';

export interface LeaderboardEntry {
  id: number;
  nickname: string;
  coins: number;
  avatar: AvatarConfig;
  rank: number;
  updatedAt: Date;
}