# lib/api-zod/src/generated/types/leaderboardEntryInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AvatarConfig } from './avatarConfig';

export interface LeaderboardEntryInput {
  /** @minLength 1 */
  deviceId: string;
  /**
     * @minLength 1
     * @maxLength 20
     */
  nickname: string;
  /** @minimum 0 */
  coins: number;
  avatar: AvatarConfig;
}